api 수정

See merge request s09-final/S09P31B309!70

+++ deathnote/src/main/java/com/goat/deathnote/redis/service/RankingService.java
@@ -21,7 +21,7 @@ public class RankingService {
         redisTemplate.opsForZSet().add(RANKING_KEY, member.getId().toString(), score);
     }
 
-    public Set<ZSetOperations.TypedTuple<String>> getTopMembers(int count) {
+    public Set<ZSetOperations.TypedTuple<String>> getTopMembers(Long count) {
         return redisTemplate.opsForZSet().reverseRangeWithScores(RANKING_KEY, 0, count - 1);
     }
 }
